% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assessment_vals.R
\name{get_assessment_vals}
\alias{get_assessment_vals}
\title{Get assessment values}
\usage{
get_assessment_vals(species = NULL, itis = NULL, year = NULL)
}
\arguments{
\item{species}{A character string of the species name}

\item{itis}{A numeric value representing the ITIS Taxon Serial Number}

\item{year}{A numeric value of the assement year of interest}
}
\value{
a data frame of SSB ratios of each species (n x 27)
\item{stock_id}{NEFSC Stock ID}
\item{stock_name}{Full name of stock}
\item{jurisdiction}{Fishery Management Council responsible for management}
\item{fmp}{The FMP the stock is managed under}
\item{assessment_year}{Most recent assessment year}
\item{last_data_year}{The modst recent year of data used as input into the assessment}
\item{f_year}{Year of F Estimate}
\item{f_unit}{Unit of measure for F}
\item{f_basis}{Basis of F estimate}
\item{flimit}{The recommended fishing mortality limit from the assessment}
\item{flimt_basis}{Basis for F_limit}
\item{fmsy}{Estimated F that would produce MSY from a stock at B_MSY}
\item{fmsy_basis}{Basis of estimate of F_MSY}
\item{b_year}{Year of the biomass estimate}
\item{estimated_b}{Estimate of current biomass from the final, accepted assessment}
\item{b_unit}{Unit of measure for the biomass estimate}
\item{b_basis}{Basis for the biomass estimate; retrospective pattern adjustments are noted here}
\item{blimit}{The recommended biomass limit from the assessment}
\item{blimit_basis}{Basis for the recommended biomass limit}
\item{bmsy}{Estimated stock size that would produce MSY when fished at F_MSY}
\item{bmsy_basis}{Basis for the estimated B_MSY}
\item{b_bmsy}{Ratio of estimated SSB to SSB_MSY}
\item{msy}{Estimate of maximum sustainable yield}
\item{msy_unit}{Unit of measure for MSY}
\item{overfished}{If B_limit < B_MSY, stock is not considered overfished}
\item{overfishing}{If F_limit < F_MSY, stock is not considered to be experiencing overfishing}
}
\description{
A function to extract the F and SSB values from the most recent stock assessment relative to the estimated SSB at MSY for NEFMC managed species.
}
\examples{
species <- "Atlantic cod"
year <- 2024 
get_assessment_vals(species = species, year = year)

itis <- 164712
get_assessment_vals(itis = itis, year = year)
}
