% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_code.R
\name{get_species_code}
\alias{get_species_code}
\title{Get species code}
\usage{
get_species_code(species = NULL)
}
\arguments{
\item{species}{A character string of the species name}
}
\value{
a data frame of species identification information (n x 3)
\item{itis_code}{ITIS Taxonomic Serial Number}
\item{species}{Full name of stock}
\item{FMP}{Fishery Management Council responsible for management}
}
\description{
A function to extract the species code or species name that is used within the GARFO Catch Accounting and Monitoring System database.
}
\examples{
species <- "Atlantic cod"
get_species_code(species = species)


get_species_code(species = "hake")
}
