% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ssb_ratio.R
\name{get_ssb_ratio}
\alias{get_ssb_ratio}
\title{Get SSB Ratio}
\usage{
get_ssb_ratio(species = NULL, itis = NULL, year = NULL)
}
\arguments{
\item{species}{A character string of the species name}

\item{itis}{A numeric value representing the ITIS Taxon Serial Number}

\item{year}{A numeric value of the assement year of interest}
}
\value{
a dataframe of SSB ratios of each species
}
\description{
A function to extract the ratio of the estimated SSB from the most recent stock assessment relative to the estimated SSB at MSY for NEFMC managed species.
}
\examples{
species <- "Atlantic cod"
year <- 2024 
get_ssb_ratio(species = species, year = year)

itis <- 164712
get_ssb_ratio(itis = itis, year = year)
}
