% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nefmc_species}
\alias{nefmc_species}
\title{NEFMC Species data}
\format{
\subsection{\code{nefmc_species}}{

A data frame with 34 rows and 16 columns:
\describe{
\item{ITIS_TSN}{Integrated Taxonomic Information System (ITIS) - Taxonomic Species Number (TSN). NULL for No-Catch records.}
\item{ITIS_SCI_NAME}{Integrated Taxonomic Information System Scientific Name}
\item{ITIS_NAME}{Integrated Taxonomic Information System species common name}
\item{ITIS_GROUP0}{General grouping of ITIS TSN for apportionment}
\item{ITIS_GROUP2}{Specific grouping of ITIS TSN}
\item{ITIS_GROUP1}{Actual grouping of ITIS TSN for apportionment}
\item{DLR_NESPP3}{The Northeast Species code, without the market category, i.e. the first 3 digits of NESPP4. Legacy code. For 2020+ use ITIS_TSN.}
\item{DLR_SPPNAME}{Northeast Species CFDBS dealer (DLR) species common name}
\item{VTR_SPPCODE}{Northeast Vessel Trip Report (VTR) species code}
\item{VTR_SPPNAME}{Northeast Vessel Trip Report (VTR) species name}
\item{FAO_3ACODE}{}
\item{FAO_SCI_NAME}{}
\item{FAO_NAME}{}
\item{FMP}{Fishery Management Plan}
\item{COUNCIL}{Council responsible for management under a federal fisheries management plan. For those species jointly managed by the New England Fisheries Management Council and the Mid-Atlantic Management Council; the first council listed is the lead council for that fisheries management plan.}
\item{ESA_LISTED}{Indicates whether a species is federally listed under the Endangered Species Act.}
}
}
}
\source{
\url{https://apps-garfo.fisheries.noaa.gov/cams}
}
\usage{
nefmc_species
}
\description{
A dataframe containing the CAMS-specific codes for each species managed by the NEFMC.
}
\keyword{datasets}
