% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_fy.R
\name{assign_fy}
\alias{assign_fy}
\title{Assign Fishing Year column}
\usage{
assign_fy(data, col, species_data, impose.fy = NULL)
}
\arguments{
\item{data}{a dataframe containing trip data for a given species}

\item{col}{the column containing the date a trip occurred}

\item{species_data}{a dataframe of species containing fishery management plan and fishing year information}

\item{impose.fy}{if the data is for a species that is not currently managed by the NEFMC, should a fishing year be imposed for the purposes of analysis. If so, which fishery management plan should be used to impose a fishing year perid.}
}
\value{
a column in the user provided dataframe that contains values for the fishing year
}
\description{
Assigns the fishing year in which a trip was recorded for a given species based on the calendar year date and the fishing year intervals for a Fishery Management Plan
}
